requirements.txt
pytest
requests
requests-html
pyaml
jinja2